// apps/backend/src/transport/third-party/revenuecat/revenuecat-api-types.ts
export interface components {
  schemas: {
// ...
    ListSubscriptions: {
      /**
       * @description String representing the object's type. Objects of the same type share the same value. Always has the value `list`.
       * @enum {string}
       */
      object: 'list'
      /** @description Details about each object. */
      items: components['schemas']['Subscription'][]
      /**
       * @description URL to access the next page of the customer's subscriptions. If not present / null, there is no next page
       * @example /v2/projects/proj1ab2c3d4/customers/19b8de26-77c1-49f1-aa18-019a391603e2/subscriptions?starting_after=sub1a2b3c4d
       */
      next_page: string | null
      /**
       * @description The URL where this list can be accessed.
       * @example /v2/projects/proj1ab2c3d4/customers/19b8de26-77c1-49f1-aa18-019a391603e2/subscriptions
       */
      url: string
    }
// ...
  }
// ...
}